move test to new file
